feature: rewrite of producer
