+++ aether-producer/manage.py
@@ -23,15 +23,12 @@ import sys
 from time import sleep
 
 if __name__ == '__main__':
-    from producer import new_producer
-    new_producer.main()
 
-    '''
     if len(sys.argv) > 1:
         if sys.argv[1].strip() == "test":
             print("starting producer for test")
             from producer import aether_producer
-            aether_producer.main_loop(test=True)
+            aether_producer.main(test=True)
             print("Started Producer for Test")
         else:
             print ("invalid argument: %s" % sys.argv[1])
@@ -39,6 +36,5 @@ if __name__ == '__main__':
         print("starting producer")
         print("waiting for requisites")
         from producer import aether_producer
-        aether_producer.main_loop()
+        aether_producer.main(test=False)
         print("Started Producer")
-    '''
