+++ aether-producer/producer/settings.json
@@ -2,9 +2,7 @@
     "offset_db_url" : "sqlite:///offset.db",
     "start_delay": 5,
     "sleep_time": 10,
-    "log_level" : "DEBUG",
-    "jdbc_connection_string": "jdbc:postgresql://db:5432/aether?user=postgres",
-    "jdbc_user": "postgres",
+    "log_level" : "INFO",
     "offset_file": "offset.json",
     "postgres_pull_limit" : 250,
     "postgres_connection_info": {
